arreglos varios

+++ db/migrate/20160319194255_create_sueldos.rb
@@ -1,10 +1,10 @@
 class CreateSueldos < ActiveRecord::Migration
   def change
     create_table :sueldos do |t|
-      t.money :monto
+      t.float :monto
       t.boolean :activo, :default =>true
       t.references :cargo, index: true, foreign_key: true
-      t.money :sueldo_integral
+      t.float :sueldo_integral
       t.timestamps null: false
     end
   end
